Restructure tests/ into directories
Mirrors the package layout and separates unit, tool, and integration
tests. Cross-test imports updated to reflect new paths.
