+++ tests/comparison/test_difference.py
