+++ tests/optimization/optimization_problem_fixtures.py
