+++ tests/optimization/test_cache.py
