+++ tests/optimization/test_pymoo.py
@@ -4,7 +4,7 @@ import unittest
 from CADETProcess import settings
 from CADETProcess.optimization import U_NSGA3
 
-from tests.test_optimization_problem import setup_optimization_problem
+from tests.optimization.test_optimization_problem import setup_optimization_problem
 
 
 class Test_OptimizationProblemSimple(unittest.TestCase):
